+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Confirmation Dialog/ConfirmationDialogParameters.swift
@@ -1,0 +1,106 @@
+//
+//  ConfirmationDialogParameters.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 10.08.22.
+//
+
+import SwiftUI
+
+// MARK: - ConfirmationDialog Parameters
+/// Confirmation dialog Parameters.
+///
+/// Parameters for presenting a `ConfirmationDialog`.
+///
+/// In `MVP`, `VIP`, and `VIPER` architectures, parameters are stored in`Presenter`.
+/// in `MVVM` architecture, parameters are stored in `ViewModel.`
+///
+///     @State private var parameters: ConfirmationDialogParameters?
+///
+///     var body: some View {
+///         Button("Lorem ipsum", action: {
+///             parameters = ConfirmationDialogParameters(
+///                 title: "Lorem Ipsum",
+///                 message: "Lorem ipsum",
+///                 actions: {
+///                     ConfirmationDialogButton(title: "Confirm", action: { print("Confirmed") })
+///                     ConfirmationDialogButton(role: .cancel, title: "Cancel", action: { print("Cancelled") })
+///                 }
+///             )
+///         })
+///             .confirmationDialog(parameters: $parameters)
+///     }
+///
+@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+public struct ConfirmationDialogParameters {
+    // MARK: Properties
+    /// Confirmation dialog title.
+    public var title: String?
+    
+    /// Confirmation dialog message.
+    public var message: String?
+    
+    /// Buttons.
+    public var buttons: () -> [any ConfirmationDialogButtonProtocol]
+    
+    // MARK: Initializers
+    /// Initializes `ConfirmationDialogParameters`.
+    public init(
+        title: String?,
+        message: String?,
+        @ConfirmationDialogButtonBuilder actions buttons: @escaping () -> [any ConfirmationDialogButtonProtocol]
+    ) {
+        self.title = title
+        self.message = message
+        self.buttons = buttons
+    }
+}
+
+// MARK: - Factory
+@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+extension View {
+    /// Presents `ConfirmationDialog` when `ConfirmationDialogParameters` is non-`nil`.
+    @ViewBuilder public func confirmationDialog(
+        parameters: Binding<ConfirmationDialogParameters?>
+    ) -> some View {
+        switch parameters.wrappedValue {
+        case nil:
+            self
+
+        case let _parameters?:
+            self.confirmationDialog(
+                _parameters.title ?? "",
+                isPresented: .constant(true),
+                titleVisibility: .confirmationDialog(title: _parameters.title, message: _parameters.message),
+                actions: {
+                    ForEach(_parameters.buttons().enumeratedArray(), id: \.offset, content: { (_, button) in
+                        button.body(
+                            animateOut: {
+                                parameters.wrappedValue = nil
+                                $0?()
+                            }
+                        )
+                    })
+                },
+                message: {
+                    if let message: String = _parameters.message {
+                        Text(message)
+                    }
+                }
+            )
+        }
+    }
+}
+
+// MARK: - Helpers
+@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+extension Visibility {
+    fileprivate static func confirmationDialog(title: String?, message: String?) -> Self {
+        switch (title, message) {
+        case (nil, nil): return .hidden
+        case (nil, _?): return .visible
+        case (_?, nil): return .visible
+        case (_?, _?): return .visible
+        }
+    }
+}